Created StyleSheet
created the style sheet and moved the styles from the screens over to StyleSheet.js
